requirements.txt
greenlet==2.0.2
numpy==1.24.3
pandas==2.0.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
tzdata==2023.3
